Use Sessions dictionary for timetable generation
All the information about a session, i.e title, matrix room id, extra
text and the talks are stored in a Session object.

Since the Sessions dictionary uses slot ID as the key, it can have
multiple sessions with the same title.

Use the Sessions dictionary for generation of the timetable and get all
the necessary information from it rather than from ad-hoc dicts.